app/requirements.txt
flask
vk_api
PyYaml
pydantic